requirements.txt
accelerate==0.21.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
airportsdata==20250523
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
astor==0.8.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
bitsandbytes==0.41.0
blake3==1.0.5
cachetools==5.5.2
captum==0.8.0
cbor2==5.6.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.2
compressed-tensors==0.10.2
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
datasets==4.0.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.6.1
einops-exts==0.0.4
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.2.3
genson==1.3.0
gguf==0.17.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grad-cam==1.4.0
gradio==3.35.2
gradio_client==0.2.9
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h5py==3.11.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
imageio==2.35.1
importlib_metadata==8.7.0
interegular==0.3.3
iso3166==2.1.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpath-ng==1.7.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
latex2mathml==3.78.0
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
llguidance==0.7.29
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==2.2.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistral_common==1.8.3
mistralai==1.8.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
open_clip_torch==2.32.0
openai==1.90.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.4.0
pillow==11.2.1
ply==3.11
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.8
ray==2.47.0
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.7.0
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.26
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.4
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
ttach==0.0.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.30.6
uvloop==0.21.0
vllm==0.9.2
watchfiles==1.1.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.0
zhipuai==2.1.5.20250626
zipp==3.23.0
